+++ modules/consensus/accept_txntypes_test.go
@@ -699,3 +699,773 @@ func TestFileContractsBlocks(t *testing.T) {
 	}
 }
 */
+
+// testSpendSiafundsBlock mines a block with a transaction spending siafunds
+// and adds it to the consensus set.
+/*
+func (cst *consensusSetTester) testSpendSiafundsBlock() error {
+	// Create a destination for the siafunds.
+	var destAddr types.UnlockHash
+	_, err := rand.Read(destAddr[:])
+	if err != nil {
+		return err
+	}
+
+	// Find the siafund output that is 'anyone can spend' (output exists only
+	// in the testing setup).
+	var srcID types.SiafundOutputID
+	var srcValue types.Currency
+	anyoneSpends := types.UnlockConditions{}.UnlockHash()
+	cst.cs.db.forEachSiafundOutputs(func(id types.SiafundOutputID, sfo types.SiafundOutput) {
+		if sfo.UnlockHash == anyoneSpends {
+			srcID = id
+			srcValue = sfo.Value
+		}
+	})
+
+	// Create a transaction that spends siafunds.
+	txn := types.Transaction{
+		SiafundInputs: []types.SiafundInput{{
+			ParentID:         srcID,
+			UnlockConditions: types.UnlockConditions{},
+		}},
+		SiafundOutputs: []types.SiafundOutput{
+			{
+				Value:      srcValue.Sub(types.NewCurrency64(1)),
+				UnlockHash: types.UnlockConditions{}.UnlockHash(),
+			},
+			{
+				Value:      types.NewCurrency64(1),
+				UnlockHash: destAddr,
+			},
+		},
+	}
+	sfoid0 := txn.SiafundOutputID(0)
+	sfoid1 := txn.SiafundOutputID(1)
+	cst.tpool.AcceptTransactionSet([]types.Transaction{txn})
+
+	// Mine a block containing the txn.
+	_, err = cst.miner.AddBlock()
+	if err != nil {
+		return err
+	}
+
+	// Check that the input got consumed, and that the outputs got created.
+	exists := cst.cs.db.inSiafundOutputs(srcID)
+	if exists {
+		return errors.New("siafund output was not properly consumed")
+	}
+	exists = cst.cs.db.inSiafundOutputs(sfoid0)
+	if !exists {
+		return errors.New("siafund output was not properly created")
+	}
+	sfo, err := cst.cs.dbGetSiafundOutput(sfoid0)
+	if err != nil {
+		return err
+	}
+	if sfo.Value.Cmp(srcValue.Sub(types.NewCurrency64(1))) != 0 {
+		return errors.New("created siafund has wrong value")
+	}
+	if sfo.UnlockHash != anyoneSpends {
+		return errors.New("siafund output sent to wrong unlock hash")
+	}
+	exists = cst.cs.db.inSiafundOutputs(sfoid1)
+	if !exists {
+		return errors.New("second siafund output was not properly created")
+	}
+	sfo, err = cst.cs.dbGetSiafundOutput(sfoid1)
+	if err != nil {
+		return err
+	}
+	if sfo.Value.Cmp(types.NewCurrency64(1)) != 0 {
+		return errors.New("second siafund output has wrong value")
+	}
+	if sfo.UnlockHash != destAddr {
+		return errors.New("second siafund output sent to wrong addr")
+	}
+
+	// Put a file contract into the blockchain that will add values to siafund
+	// outputs.
+	var siafundPool types.Currency
+	err = cst.cs.db.Update(func(tx *bolt.Tx) error {
+		siafundPool = getSiafundPool(tx)
+		return nil
+	})
+	if err != nil {
+		panic(err)
+	}
+	oldSiafundPool := siafundPool
+	payout := types.NewCurrency64(400e6)
+	fc := types.FileContract{
+		WindowStart: cst.cs.dbBlockHeight() + 2,
+		WindowEnd:   cst.cs.dbBlockHeight() + 4,
+		Payout:      payout,
+	}
+	outputSize := payout.Sub(types.Tax(cst.cs.dbBlockHeight(), fc.Payout))
+	fc.ValidProofOutputs = []types.SiacoinOutput{{Value: outputSize}}
+	fc.MissedProofOutputs = []types.SiacoinOutput{{Value: outputSize}}
+
+	// Create and fund a transaction with a file contract.
+	txnBuilder := cst.wallet.StartTransaction()
+	err = txnBuilder.FundSiacoins(payout)
+	if err != nil {
+		return err
+	}
+	txnBuilder.AddFileContract(fc)
+	txnSet, err := txnBuilder.Sign(true)
+	if err != nil {
+		return err
+	}
+	err = cst.tpool.AcceptTransactionSet(txnSet)
+	if err != nil {
+		return err
+	}
+	_, err = cst.miner.AddBlock()
+	if err != nil {
+		return err
+	}
+	err = cst.cs.db.Update(func(tx *bolt.Tx) error {
+		siafundPool = getSiafundPool(tx)
+		return nil
+	})
+	if err != nil {
+		panic(err)
+	}
+	if siafundPool.Cmp(types.NewCurrency64(15600e3).Add(oldSiafundPool)) != 0 {
+		return errors.New("siafund pool did not update correctly")
+	}
+
+	// Create a transaction that spends siafunds.
+	var claimDest types.UnlockHash
+	_, err = rand.Read(claimDest[:])
+	if err != nil {
+		return err
+	}
+	var srcClaimStart types.Currency
+	cst.cs.db.forEachSiafundOutputs(func(id types.SiafundOutputID, sfo types.SiafundOutput) {
+		if sfo.UnlockHash == anyoneSpends {
+			srcID = id
+			srcValue = sfo.Value
+			srcClaimStart = sfo.ClaimStart
+		}
+	})
+	txn = types.Transaction{
+		SiafundInputs: []types.SiafundInput{{
+			ParentID:         srcID,
+			UnlockConditions: types.UnlockConditions{},
+			ClaimUnlockHash:  claimDest,
+		}},
+		SiafundOutputs: []types.SiafundOutput{
+			{
+				Value:      srcValue.Sub(types.NewCurrency64(1)),
+				UnlockHash: types.UnlockConditions{}.UnlockHash(),
+			},
+			{
+				Value:      types.NewCurrency64(1),
+				UnlockHash: destAddr,
+			},
+		},
+	}
+	sfoid1 = txn.SiafundOutputID(1)
+	cst.tpool.AcceptTransactionSet([]types.Transaction{txn})
+	_, err = cst.miner.AddBlock()
+	if err != nil {
+		return err
+	}
+
+	// Find the siafund output and check that it has the expected number of
+	// siafunds.
+	err = cst.cs.db.Update(func(tx *bolt.Tx) error {
+		siafundPool = getSiafundPool(tx)
+		return nil
+	})
+	if err != nil {
+		panic(err)
+	}
+	found := false
+	expectedBalance := siafundPool.Sub(srcClaimStart).Div(types.NewCurrency64(10e3)).Mul(srcValue)
+	cst.cs.db.forEachDelayedSiacoinOutputsHeight(cst.cs.dbBlockHeight()+types.MaturityDelay, func(id types.SiacoinOutputID, output types.SiacoinOutput) {
+		if output.UnlockHash == claimDest {
+			found = true
+			if output.Value.Cmp(expectedBalance) != 0 {
+				// err is scoped outside this func
+				err = errors.New("siafund output has the wrong balance")
+			}
+		}
+	})
+	if err != nil {
+		return err
+	}
+	if !found {
+		return errors.New("could not find siafund claim output")
+	}
+
+	return nil
+}
+*/
+
+// TestSpendSiafundsBlock creates a consensus set tester and uses it to call
+// testSpendSiafundsBlock.
+/*
+func TestSpendSiafundsBlock(t *testing.T) {
+	if testing.Short() {
+		t.SkipNow()
+	}
+	cst, err := createConsensusSetTester("TestSpendSiafundsBlock")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer cst.closeCst()
+
+	// COMPATv0.4.0
+	//
+	// Mine enough blocks to get above the file contract hardfork threshold
+	// (10).
+	for i := 0; i < 10; i++ {
+		_, err = cst.miner.AddBlock()
+		if err != nil {
+			t.Fatal(err)
+		}
+	}
+	err = cst.testSpendSiafundsBlock()
+	if err != nil {
+		t.Error(err)
+	}
+}
+*/
+
+// testPaymentChannelBlocks submits blocks to set up, use, and close a payment
+// channel.
+/*
+func (cst *consensusSetTester) testPaymentChannelBlocks() error {
+	// The current method of doing payment channels is gimped because public
+	// keys do not have timelocks. We will be hardforking to include timelocks
+	// in public keys in 0.4.0, but in the meantime we need an alternate
+	// method.
+
+	// Gimped payment channels: 2-of-2 multisig where one key is controlled by
+	// the funding entity, and one key is controlled by the receiving entity. An
+	// address is created containing both keys, and then the funding entity
+	// creates, but does not sign, a transaction sending coins to the channel
+	// address. A second transaction is created that sends all the coins in the
+	// funding output back to the funding entity. The receiving entity signs the
+	// transaction with a timelocked signature. The funding entity will get the
+	// refund after T blocks as long as the output is not double spent. The
+	// funding entity then signs the first transaction and opens the channel.
+	//
+	// Creating the channel:
+	//	1. Create a 2-of-2 unlock conditions, one key held by each entity.
+	//	2. Funding entity creates, but does not sign, a transaction sending
+	//		money to the payment channel address. (txn A)
+	//	3. Funding entity creates and signs a transaction spending the output
+	//		created in txn A that sends all the money back as a refund. (txn B)
+	//	4. Receiving entity signs txn B with a timelocked signature, so that the
+	//		funding entity cannot get the refund for several days. The funding entity
+	//		is given a fully signed and eventually-spendable txn B.
+	//	5. The funding entity signs and broadcasts txn A.
+	//
+	// Using the channel:
+	//	Each the receiving entity and the funding entity keeps a record of how
+	//	much has been sent down the unclosed channel, and watches the
+	//	blockchain for a channel closing transaction. To send more money down
+	//	the channel, the funding entity creates and signs a transaction sending
+	//	X+y coins to the receiving entity from the channel address. The
+	//	transaction is sent to the receiving entity, who will keep it and
+	//	potentially sign and broadcast it later. The funding entity will only
+	//	send money down the channel if 'work' or some other sort of event has
+	//	completed that indicates the receiving entity should get more money.
+	//
+	// Closing the channel:
+	//	The receiving entity will sign the transaction that pays them the most
+	//	money and then broadcast that transaction. This will spend the output
+	//	and close the channel, invalidating txn B and preventing any future
+	//	transactions from being made over the channel. The channel must be
+	//	closed before the timelock expires on the second signature in txn B,
+	//	otherwise the funding entity will be able to get a full refund.
+	//
+	//	The funding entity should be waiting until either the receiving entity
+	//	closes the channel or the timelock expires. If the receiving entity
+	//	closes the channel, all is good. If not, then the funding entity can
+	//	close the channel and get a full refund.
+
+	// Create a 2-of-2 unlock conditions, 1 key for each the sender and the
+	// receiver in the payment channel.
+	sk1, pk1, err := crypto.GenerateSignatureKeys() // Funding entity.
+	if err != nil {
+		return err
+	}
+	sk2, pk2, err := crypto.GenerateSignatureKeys() // Receiving entity.
+	if err != nil {
+		return err
+	}
+	uc := types.UnlockConditions{
+		PublicKeys: []types.SiaPublicKey{
+			{
+				Algorithm: types.SignatureEd25519,
+				Key:       pk1[:],
+			},
+			{
+				Algorithm: types.SignatureEd25519,
+				Key:       pk2[:],
+			},
+		},
+		SignaturesRequired: 2,
+	}
+	channelAddress := uc.UnlockHash()
+
+	// Funding entity creates but does not sign a transaction that funds the
+	// channel address. Because the wallet is not very flexible, the channel
+	// txn needs to be fully custom. To get a custom txn, manually create an
+	// address and then use the wallet to fund that address.
+	channelSize := types.NewCurrency64(10e3)
+	channelFundingSK, channelFundingPK, err := crypto.GenerateSignatureKeys()
+	if err != nil {
+		return err
+	}
+	channelFundingUC := types.UnlockConditions{
+		PublicKeys: []types.SiaPublicKey{{
+			Algorithm: types.SignatureEd25519,
+			Key:       channelFundingPK[:],
+		}},
+		SignaturesRequired: 1,
+	}
+	channelFundingAddr := channelFundingUC.UnlockHash()
+	fundTxnBuilder := cst.wallet.StartTransaction()
+	if err != nil {
+		return err
+	}
+	err = fundTxnBuilder.FundSiacoins(channelSize)
+	if err != nil {
+		return err
+	}
+	scoFundIndex := fundTxnBuilder.AddSiacoinOutput(types.SiacoinOutput{Value: channelSize, UnlockHash: channelFundingAddr})
+	fundTxnSet, err := fundTxnBuilder.Sign(true)
+	if err != nil {
+		return err
+	}
+	fundOutputID := fundTxnSet[len(fundTxnSet)-1].SiacoinOutputID(int(scoFundIndex))
+	channelTxn := types.Transaction{
+		SiacoinInputs: []types.SiacoinInput{{
+			ParentID:         fundOutputID,
+			UnlockConditions: channelFundingUC,
+		}},
+		SiacoinOutputs: []types.SiacoinOutput{{
+			Value:      channelSize,
+			UnlockHash: channelAddress,
+		}},
+		TransactionSignatures: []types.TransactionSignature{{
+			ParentID:       crypto.Hash(fundOutputID),
+			PublicKeyIndex: 0,
+			CoveredFields:  types.CoveredFields{WholeTransaction: true},
+		}},
+	}
+
+	// Funding entity creates and signs a transaction that spends the full
+	// channel output.
+	channelOutputID := channelTxn.SiacoinOutputID(0)
+	refundUC, err := cst.wallet.NextAddress()
+	refundAddr := refundUC.UnlockHash()
+	if err != nil {
+		return err
+	}
+	refundTxn := types.Transaction{
+		SiacoinInputs: []types.SiacoinInput{{
+			ParentID:         channelOutputID,
+			UnlockConditions: uc,
+		}},
+		SiacoinOutputs: []types.SiacoinOutput{{
+			Value:      channelSize,
+			UnlockHash: refundAddr,
+		}},
+		TransactionSignatures: []types.TransactionSignature{{
+			ParentID:       crypto.Hash(channelOutputID),
+			PublicKeyIndex: 0,
+			CoveredFields:  types.CoveredFields{WholeTransaction: true},
+		}},
+	}
+	sigHash := refundTxn.SigHash(0)
+	cryptoSig1, err := crypto.SignHash(sigHash, sk1)
+	if err != nil {
+		return err
+	}
+	refundTxn.TransactionSignatures[0].Signature = cryptoSig1[:]
+
+	// Receiving entity signs the transaction that spends the full channel
+	// output, but with a timelock.
+	refundTxn.TransactionSignatures = append(refundTxn.TransactionSignatures, types.TransactionSignature{
+		ParentID:       crypto.Hash(channelOutputID),
+		PublicKeyIndex: 1,
+		Timelock:       cst.cs.dbBlockHeight() + 2,
+		CoveredFields:  types.CoveredFields{WholeTransaction: true},
+	})
+	sigHash = refundTxn.SigHash(1)
+	cryptoSig2, err := crypto.SignHash(sigHash, sk2)
+	if err != nil {
+		return err
+	}
+	refundTxn.TransactionSignatures[1].Signature = cryptoSig2[:]
+
+	// Funding entity will now sign and broadcast the funding transaction.
+	sigHash = channelTxn.SigHash(0)
+	cryptoSig0, err := crypto.SignHash(sigHash, channelFundingSK)
+	if err != nil {
+		return err
+	}
+	channelTxn.TransactionSignatures[0].Signature = cryptoSig0[:]
+	err = cst.tpool.AcceptTransactionSet(append(fundTxnSet, channelTxn))
+	if err != nil {
+		return err
+	}
+	// Put the txn in a block.
+	_, err = cst.miner.AddBlock()
+	if err != nil {
+		return err
+	}
+
+	// Try to submit the refund transaction before the timelock has expired.
+	err = cst.tpool.AcceptTransactionSet([]types.Transaction{refundTxn})
+	if err != types.ErrPrematureSignature {
+		return err
+	}
+
+	// Create a transaction that has partially used the channel, and submit it
+	// to the blockchain to close the channel.
+	closeTxn := types.Transaction{
+		SiacoinInputs: []types.SiacoinInput{{
+			ParentID:         channelOutputID,
+			UnlockConditions: uc,
+		}},
+		SiacoinOutputs: []types.SiacoinOutput{
+			{
+				Value:      channelSize.Sub(types.NewCurrency64(5)),
+				UnlockHash: refundAddr,
+			},
+			{
+				Value: types.NewCurrency64(5),
+			},
+		},
+		TransactionSignatures: []types.TransactionSignature{
+			{
+				ParentID:       crypto.Hash(channelOutputID),
+				PublicKeyIndex: 0,
+				CoveredFields:  types.CoveredFields{WholeTransaction: true},
+			},
+			{
+				ParentID:       crypto.Hash(channelOutputID),
+				PublicKeyIndex: 1,
+				CoveredFields:  types.CoveredFields{WholeTransaction: true},
+			},
+		},
+	}
+	sigHash = closeTxn.SigHash(0)
+	cryptoSig3, err := crypto.SignHash(sigHash, sk1)
+	if err != nil {
+		return err
+	}
+	closeTxn.TransactionSignatures[0].Signature = cryptoSig3[:]
+	sigHash = closeTxn.SigHash(1)
+	cryptoSig4, err := crypto.SignHash(sigHash, sk2)
+	if err != nil {
+		return err
+	}
+	closeTxn.TransactionSignatures[1].Signature = cryptoSig4[:]
+	err = cst.tpool.AcceptTransactionSet([]types.Transaction{closeTxn})
+	if err != nil {
+		return err
+	}
+
+	// Mine the block with the transaction.
+	_, err = cst.miner.AddBlock()
+	if err != nil {
+		return err
+	}
+	closeRefundID := closeTxn.SiacoinOutputID(0)
+	closePaymentID := closeTxn.SiacoinOutputID(1)
+	exists := cst.cs.db.inSiacoinOutputs(closeRefundID)
+	if !exists {
+		return errors.New("close txn refund output doesn't exist")
+	}
+	exists = cst.cs.db.inSiacoinOutputs(closePaymentID)
+	if !exists {
+		return errors.New("close txn payment output doesn't exist")
+	}
+
+	// Create a payment channel where the receiving entity never responds to
+	// the initial transaction.
+	{
+		// Funding entity creates but does not sign a transaction that funds the
+		// channel address. Because the wallet is not very flexible, the channel
+		// txn needs to be fully custom. To get a custom txn, manually create an
+		// address and then use the wallet to fund that address.
+		channelSize := types.NewCurrency64(10e3)
+		channelFundingSK, channelFundingPK, err := crypto.GenerateSignatureKeys()
+		if err != nil {
+			return err
+		}
+		channelFundingUC := types.UnlockConditions{
+			PublicKeys: []types.SiaPublicKey{{
+				Algorithm: types.SignatureEd25519,
+				Key:       channelFundingPK[:],
+			}},
+			SignaturesRequired: 1,
+		}
+		channelFundingAddr := channelFundingUC.UnlockHash()
+		fundTxnBuilder := cst.wallet.StartTransaction()
+		err = fundTxnBuilder.FundSiacoins(channelSize)
+		if err != nil {
+			return err
+		}
+		scoFundIndex := fundTxnBuilder.AddSiacoinOutput(types.SiacoinOutput{Value: channelSize, UnlockHash: channelFundingAddr})
+		fundTxnSet, err := fundTxnBuilder.Sign(true)
+		if err != nil {
+			return err
+		}
+		fundOutputID := fundTxnSet[len(fundTxnSet)-1].SiacoinOutputID(int(scoFundIndex))
+		channelTxn := types.Transaction{
+			SiacoinInputs: []types.SiacoinInput{{
+				ParentID:         fundOutputID,
+				UnlockConditions: channelFundingUC,
+			}},
+			SiacoinOutputs: []types.SiacoinOutput{{
+				Value:      channelSize,
+				UnlockHash: channelAddress,
+			}},
+			TransactionSignatures: []types.TransactionSignature{{
+				ParentID:       crypto.Hash(fundOutputID),
+				PublicKeyIndex: 0,
+				CoveredFields:  types.CoveredFields{WholeTransaction: true},
+			}},
+		}
+
+		// Funding entity creates and signs a transaction that spends the full
+		// channel output.
+		channelOutputID := channelTxn.SiacoinOutputID(0)
+		refundUC, err := cst.wallet.NextAddress()
+		refundAddr := refundUC.UnlockHash()
+		if err != nil {
+			return err
+		}
+		refundTxn := types.Transaction{
+			SiacoinInputs: []types.SiacoinInput{{
+				ParentID:         channelOutputID,
+				UnlockConditions: uc,
+			}},
+			SiacoinOutputs: []types.SiacoinOutput{{
+				Value:      channelSize,
+				UnlockHash: refundAddr,
+			}},
+			TransactionSignatures: []types.TransactionSignature{{
+				ParentID:       crypto.Hash(channelOutputID),
+				PublicKeyIndex: 0,
+				CoveredFields:  types.CoveredFields{WholeTransaction: true},
+			}},
+		}
+		sigHash := refundTxn.SigHash(0)
+		cryptoSig1, err := crypto.SignHash(sigHash, sk1)
+		if err != nil {
+			return err
+		}
+		refundTxn.TransactionSignatures[0].Signature = cryptoSig1[:]
+
+		// Recieving entity never communitcates, funding entity must reclaim
+		// the 'channelSize' coins that were intended to go to the channel.
+		reclaimUC, err := cst.wallet.NextAddress()
+		reclaimAddr := reclaimUC.UnlockHash()
+		if err != nil {
+			return err
+		}
+		reclaimTxn := types.Transaction{
+			SiacoinInputs: []types.SiacoinInput{{
+				ParentID:         fundOutputID,
+				UnlockConditions: channelFundingUC,
+			}},
+			SiacoinOutputs: []types.SiacoinOutput{{
+				Value:      channelSize,
+				UnlockHash: reclaimAddr,
+			}},
+			TransactionSignatures: []types.TransactionSignature{{
+				ParentID:       crypto.Hash(fundOutputID),
+				PublicKeyIndex: 0,
+				CoveredFields:  types.CoveredFields{WholeTransaction: true},
+			}},
+		}
+		sigHash = reclaimTxn.SigHash(0)
+		cryptoSig, err := crypto.SignHash(sigHash, channelFundingSK)
+		if err != nil {
+			return err
+		}
+		reclaimTxn.TransactionSignatures[0].Signature = cryptoSig[:]
+		err = cst.tpool.AcceptTransactionSet(append(fundTxnSet, reclaimTxn))
+		if err != nil {
+			return err
+		}
+		block, _ := cst.miner.FindBlock()
+		err = cst.cs.AcceptBlock(block)
+		if err != nil {
+			return err
+		}
+		reclaimOutputID := reclaimTxn.SiacoinOutputID(0)
+		exists := cst.cs.db.inSiacoinOutputs(reclaimOutputID)
+		if !exists {
+			return errors.New("failed to reclaim an output that belongs to the funding entity")
+		}
+	}
+
+	// Create a channel and the open the channel, but close the channel using
+	// the timelocked signature.
+	{
+		// Funding entity creates but does not sign a transaction that funds the
+		// channel address. Because the wallet is not very flexible, the channel
+		// txn needs to be fully custom. To get a custom txn, manually create an
+		// address and then use the wallet to fund that address.
+		channelSize := types.NewCurrency64(10e3)
+		channelFundingSK, channelFundingPK, err := crypto.GenerateSignatureKeys()
+		if err != nil {
+			return err
+		}
+		channelFundingUC := types.UnlockConditions{
+			PublicKeys: []types.SiaPublicKey{{
+				Algorithm: types.SignatureEd25519,
+				Key:       channelFundingPK[:],
+			}},
+			SignaturesRequired: 1,
+		}
+		channelFundingAddr := channelFundingUC.UnlockHash()
+		fundTxnBuilder := cst.wallet.StartTransaction()
+		err = fundTxnBuilder.FundSiacoins(channelSize)
+		if err != nil {
+			return err
+		}
+		scoFundIndex := fundTxnBuilder.AddSiacoinOutput(types.SiacoinOutput{Value: channelSize, UnlockHash: channelFundingAddr})
+		fundTxnSet, err := fundTxnBuilder.Sign(true)
+		if err != nil {
+			return err
+		}
+		fundOutputID := fundTxnSet[len(fundTxnSet)-1].SiacoinOutputID(int(scoFundIndex))
+		channelTxn := types.Transaction{
+			SiacoinInputs: []types.SiacoinInput{{
+				ParentID:         fundOutputID,
+				UnlockConditions: channelFundingUC,
+			}},
+			SiacoinOutputs: []types.SiacoinOutput{{
+				Value:      channelSize,
+				UnlockHash: channelAddress,
+			}},
+			TransactionSignatures: []types.TransactionSignature{{
+				ParentID:       crypto.Hash(fundOutputID),
+				PublicKeyIndex: 0,
+				CoveredFields:  types.CoveredFields{WholeTransaction: true},
+			}},
+		}
+
+		// Funding entity creates and signs a transaction that spends the full
+		// channel output.
+		channelOutputID := channelTxn.SiacoinOutputID(0)
+		refundUC, err := cst.wallet.NextAddress()
+		refundAddr := refundUC.UnlockHash()
+		if err != nil {
+			return err
+		}
+		refundTxn := types.Transaction{
+			SiacoinInputs: []types.SiacoinInput{{
+				ParentID:         channelOutputID,
+				UnlockConditions: uc,
+			}},
+			SiacoinOutputs: []types.SiacoinOutput{{
+				Value:      channelSize,
+				UnlockHash: refundAddr,
+			}},
+			TransactionSignatures: []types.TransactionSignature{{
+				ParentID:       crypto.Hash(channelOutputID),
+				PublicKeyIndex: 0,
+				CoveredFields:  types.CoveredFields{WholeTransaction: true},
+			}},
+		}
+		sigHash := refundTxn.SigHash(0)
+		cryptoSig1, err := crypto.SignHash(sigHash, sk1)
+		if err != nil {
+			return err
+		}
+		refundTxn.TransactionSignatures[0].Signature = cryptoSig1[:]
+
+		// Receiving entity signs the transaction that spends the full channel
+		// output, but with a timelock.
+		refundTxn.TransactionSignatures = append(refundTxn.TransactionSignatures, types.TransactionSignature{
+			ParentID:       crypto.Hash(channelOutputID),
+			PublicKeyIndex: 1,
+			Timelock:       cst.cs.dbBlockHeight() + 2,
+			CoveredFields:  types.CoveredFields{WholeTransaction: true},
+		})
+		sigHash = refundTxn.SigHash(1)
+		cryptoSig2, err := crypto.SignHash(sigHash, sk2)
+		if err != nil {
+			return err
+		}
+		refundTxn.TransactionSignatures[1].Signature = cryptoSig2[:]
+
+		// Funding entity will now sign and broadcast the funding transaction.
+		sigHash = channelTxn.SigHash(0)
+		cryptoSig0, err := crypto.SignHash(sigHash, channelFundingSK)
+		if err != nil {
+			return err
+		}
+		channelTxn.TransactionSignatures[0].Signature = cryptoSig0[:]
+		err = cst.tpool.AcceptTransactionSet(append(fundTxnSet, channelTxn))
+		if err != nil {
+			return err
+		}
+		// Put the txn in a block.
+		block, _ := cst.miner.FindBlock()
+		err = cst.cs.AcceptBlock(block)
+		if err != nil {
+			return err
+		}
+
+		// Receiving entity never signs another transaction, so the funding
+		// entity waits until the timelock is complete, and then submits the
+		// refundTxn.
+		for i := 0; i < 3; i++ {
+			block, _ := cst.miner.FindBlock()
+			err = cst.cs.AcceptBlock(block)
+			if err != nil {
+				return err
+			}
+		}
+		err = cst.tpool.AcceptTransactionSet([]types.Transaction{refundTxn})
+		if err != nil {
+			return err
+		}
+		block, _ = cst.miner.FindBlock()
+		err = cst.cs.AcceptBlock(block)
+		if err != nil {
+			return err
+		}
+		refundOutputID := refundTxn.SiacoinOutputID(0)
+		exists := cst.cs.db.inSiacoinOutputs(refundOutputID)
+		if !exists {
+			return errors.New("timelocked refund transaction did not get spent correctly")
+		}
+	}
+
+	return nil
+}
+*/
+
+// TestPaymentChannelBlocks creates a consensus set tester and uses it to call
+// testPaymentChannelBlocks.
+/*
+func TestPaymentChannelBlocks(t *testing.T) {
+	if testing.Short() {
+		t.SkipNow()
+	}
+	cst, err := createConsensusSetTester("TestPaymentChannelBlocks")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer cst.closeCst()
+	err = cst.testPaymentChannelBlocks()
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+*/
